% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bgg.prepare.data.R
\name{bgg.prepare.data}
\alias{bgg.prepare.data}
\title{Prepare the dataset for the mining stuff}
\usage{
bgg.prepare.data(bgg.dataset)
}
\arguments{
\item{bgg.dataset}{the BoardGames dataset object}
}
\description{
This function will prepare the dataset removing unuseful character attributes,
removing videogames (and videogames-only attributes).
It also remove the 'details.name' attribute and the 'stats.median' column,
as it seems to be broken in BGG XMLAPI2.
Some exploratory analysis bring us that lot of games have yearpublished=0:
this is definitively a missing we put to NA.
Then, we removed games with yearpublished < 1959 (in the .05 percentile and really unuseful)
and > 2015 (games not released yet)
}

