% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgg.prepare.dummy.R
\name{bgg.prepare.dummy}
\alias{bgg.prepare.dummy}
\title{Decode dummy variables}
\usage{
bgg.prepare.dummy(bgg.dataset)
}
\arguments{
\item{bgg.dataset}{the BoardGames dataset object}
}
\description{
This function will create the dummy variables for categories, mechanics, families and implementations.
It doesn't drop the original columns.
}
