% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bgg.prepare.data.R
\name{bgg.prepare.data}
\alias{bgg.prepare.data}
\title{Prepare the dataset for the mining stuff}
\usage{
bgg.prepare.data(bgg.dataset)
}
\arguments{
\item{bgg.dataset}{the BoardGames dataset object}
}
\description{
This function will prepare the dataset removing unuseful character attributes,
removing videogames (and videogames-only attributes) and finally creating the dummyvars
for boardgame category and mechanics.
It also remove the 'details.name' attribute (you can always merge with the BoardGame dataset to retrive it)
and the 'stats.median' column, as it seems to be broken in BGG XMLAPI2.
}

