% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bgg.prepare.dummy.R
\name{bgg.prepare.dummy}
\alias{bgg.prepare.dummy}
\title{Prepare the dataset for the mining stuff}
\usage{
bgg.prepare.dummy(bgg.dataset)
}
\arguments{
\item{bgg.dataset}{the BoardGames dataset object}
}
\description{
This function will create the dummy variables from category and mechanic
}

