% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BoardGames}
\alias{BoardGames}
\title{The complete BoardGameGeeks dataset (as October 2015)}
\format{A data frame with 79916 rows and 97 variables}
\source{
\url{https://boardgamegeek.com/}
}
\usage{
data(BoardGames)
}
\description{
A dataset containing the attributes and the ratings for around 80,000 boardgames
in a format suitable for data mining.
}
\keyword{datasets}

